src/django-app/requirements.txt
appnope==0.1.4
arrow==1.4.0
asgiref==3.11.1
asttokens==3.0.1
binaryornot==0.6.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
comm==0.2.3
cookiecutter==2.7.1
debugpy==1.8.20
decorator==5.2.1
Django==6.0.3
django-browser-reload==1.21.0
django-tailwind==4.4.2
djangorestframework==3.17.0
executing==2.2.1
honcho==2.0.0
idna==3.11
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
platformdirs==4.9.4
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.3.3
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
text-unidecode==1.3
tornado==6.5.5
traitlets==5.14.3
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
folium==0.20.0  